Moved some of the impls and reordered imports, using super where possible
