+++ backend/src/authentication/error.rs
@@ -1,5 +1,3 @@
-use crate::APIError;
-
 #[derive(Debug)]
 pub enum AuthenticationError {
     UserNotFound,
@@ -25,11 +23,5 @@ impl From<password_hash::Error> for AuthenticationError {
 impl From<sqlx::Error> for AuthenticationError {
     fn from(err: sqlx::Error) -> Self {
         AuthenticationError::Database(err)
-    }
-}
-
-impl From<AuthenticationError> for APIError {
-    fn from(err: AuthenticationError) -> Self {
-        APIError::Authentication(err)
     }
 }
