+++ backend/src/authentication/password.rs
@@ -1,11 +1,10 @@
-use std::ops::Deref;
-
 use argon2::{
     Algorithm, Argon2, Params, Version,
     password_hash::{PasswordHash, PasswordHasher, PasswordVerifier, SaltString, rand_core::OsRng},
 };
 use serde::Deserialize;
 use sqlx::Type;
+use std::ops::Deref;
 
 use super::error::AuthenticationError;
