+++ backend/src/authentication/session.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 use axum::{
     extract::{FromRef, OptionalFromRequestParts},
     http::request::Parts,
@@ -9,14 +7,14 @@ use chrono::{DateTime, TimeDelta, Utc};
 use cookie::CookieBuilder;
 use serde::{Deserialize, Serialize};
 use sqlx::{FromRow, SqlitePool};
-
-use crate::{APIError, AppState};
+use std::time::Duration;
 
 use super::{
     SESSION_COOKIE_NAME, SESSION_LIFE_TIME,
     error::AuthenticationError,
     util::{create_new_session_key, get_expires_at},
 };
+use crate::{APIError, AppState};
 
 /// A session object, corresponds to a row in the sessions table
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, Hash, FromRow)]
