+++ backend/src/data_entry/status.rs
@@ -1,15 +1,13 @@
-use std::fmt::Display;
-
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::Type;
+use std::fmt::Display;
 use utoipa::ToSchema;
 
-use super::{DataError, ValidationResults, validate_polling_station_results};
-use crate::{
-    APIError, data_entry::PollingStationResults, election::Election, error::ErrorReference,
-    polling_station::PollingStation,
+use super::{
+    DataError, PollingStationResults, ValidationResults, validate_polling_station_results,
 };
+use crate::{election::Election, polling_station::PollingStation};
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum DataEntryTransitionError {
@@ -35,22 +33,6 @@ impl From<DataError> for DataEntryTransitionError {
 impl From<ValidationResults> for DataEntryTransitionError {
     fn from(err: ValidationResults) -> Self {
         Self::ValidationError(err)
-    }
-}
-
-impl From<DataEntryTransitionError> for APIError {
-    fn from(err: DataEntryTransitionError) -> Self {
-        match err {
-            DataEntryTransitionError::FirstEntryAlreadyClaimed
-            | DataEntryTransitionError::SecondEntryAlreadyClaimed => {
-                APIError::Conflict(err.to_string(), ErrorReference::DataEntryAlreadyClaimed)
-            }
-            DataEntryTransitionError::FirstEntryAlreadyFinalised
-            | DataEntryTransitionError::SecondEntryAlreadyFinalised => {
-                APIError::Conflict(err.to_string(), ErrorReference::DataEntryAlreadyFinalised)
-            }
-            _ => APIError::Conflict(err.to_string(), ErrorReference::InvalidStateTransition),
-        }
     }
 }
