+++ backend/src/data_entry/structs.rs
@@ -1,15 +1,16 @@
-use crate::{
-    APIError,
-    audit_log::DataEntryDetails,
-    data_entry::status::DataEntryStatus,
-    election::{CandidateNumber, PGNumber, PoliticalGroup},
-    error::ErrorReference,
-};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::{FromRow, types::Json};
 use std::ops::AddAssign;
 use utoipa::ToSchema;
+
+use super::status::DataEntryStatus;
+use crate::{
+    APIError,
+    audit_log::DataEntryDetails,
+    election::{CandidateNumber, PGNumber, PoliticalGroup},
+    error::ErrorReference,
+};
 
 #[derive(Serialize, Deserialize, Clone, ToSchema, Debug, FromRow, Default)]
 pub struct PollingStationDataEntry {
