+++ backend/src/report/api.rs
@@ -1,3 +1,8 @@
+use axum::extract::{Path, State};
+use axum_extra::response::Attachment;
+use utoipa_axum::{router::OpenApiRouter, routes};
+use zip::{result::ZipError, write::SimpleFileOptions};
+
 use crate::{
     APIError, AppState, ErrorResponse,
     authentication::Coordinator,
@@ -11,10 +16,6 @@ use crate::{
     polling_station::{repository::PollingStations, structs::PollingStation},
     summary::ElectionSummary,
 };
-use axum::extract::{Path, State};
-use axum_extra::response::Attachment;
-use utoipa_axum::{router::OpenApiRouter, routes};
-use zip::{result::ZipError, write::SimpleFileOptions};
 
 pub fn router() -> OpenApiRouter<AppState> {
     OpenApiRouter::default()
